cold_email_ai/requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
chromadb
beautifulsoup4
requests
tiktoken
python-dotenv
pypdf
python-docx
python-multipart
fastapi
uvicorn[standard]

